+++ 3 задание.py
@@ -1,0 +1,90 @@
+# Напишите функцию modification(lst), которая принимает список и меняет местами его первый и последний элемент.
+# В исходном списке минимум 2 элемента.
+
+def modification(lst):
+    lst[0], lst[-1] = lst[-1], lst[0]
+    # Здесь пишем код
+    return lst
+
+# Ниже НИЧЕГО НЕ НАДО ИЗМЕНЯТЬ
+
+
+
+data = [
+    [1, 2, 3],
+    [1, 2, 3, 4, 5],
+    ['н', 'л', 'о', 'с']
+]
+
+test_data = [
+    [3, 2, 1], [5, 2, 3, 4, 1], ['с', 'л', 'о', 'н']
+]
+
+
+for i, d in enumerate(data):
+    assert modification(d) == test_data[i], f'С набором {d} есть ошибка, не проходит проверку'
+    print(f'Тестовый набор {d} прошёл проверку')
+print('Всё ок')
+
+# Дан список из 7 различных элементов. Используя функции (не использовать цикл), необходимо найти:
+# минимальный и максимальный элементы списка;
+# сумму и среднее арифметическое с округлением до 2 знаков после запятой;
+
+from statistics import mean
+def get_list_info(lst):
+    max_elem = max(lst)
+    min_elem = min(lst)
+    sum_list = round(sum(lst),2)
+    average = round(mean(lst),2)
+    return min_elem, max_elem, sum_list, average
+
+# Ниже НИЧЕГО НЕ НАДО ИЗМЕНЯТЬ
+
+
+data = [
+    [1, 2, 3, 4, 5, 6, 7],
+    [-1, -2, -3, -4, -5, -6, -7],
+    [99, 56, 209, -308, -12, -18, 42],
+    [-1, -2, -3, 0, 1, 2, 3],
+]
+
+test_data = [
+    (1, 7, 28, 4.0), (-7, -1, -28, -4.0), (-308, 209, 68, 9.71), (-3, 3, 0, 0.0)
+]
+
+
+for i, d in enumerate(data):
+    assert get_list_info(d) == test_data[i], f'С набором {d} есть ошибка, не проходит проверку'
+    print(f'Тестовый набор {d} прошёл проверку')
+print('Всё ок')
+
+
+
+
+
+# Дан список. Найдите сумму элементом с четными индексами
+
+
+def even_sum(lst):
+    sum_list = sum(lst[0::2])
+    return sum_list
+
+# Ниже НИЧЕГО НЕ НАДО ИЗМЕНЯТЬ
+
+
+data = [
+    [1, 2, 3, 4, 5, 6, 7],
+    [-1, -2, -3, -4, -5, -6, -7],
+    [99, 56, 209, -308, -12, -18, 42],
+    [-1, -2, -3, 0, 1, 2, 3],
+]
+
+test_data = [
+    16, -16, 338, 0
+]
+
+
+for i, d in enumerate(data):
+    assert even_sum(d) == test_data[i], f'С набором {d} есть ошибка, не проходит проверку'
+    print(f'Тестовый набор {d} прошёл проверку')
+print('Всё ок')
